al-ai-agent-v2/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pgvector>=0.2.3

# Vector store
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0

# LangGraph and LangChain
langgraph>=1.0.0
langchain>=1.0.0
langchain-openai>=0.2.0

# LangGraph checkpointing
langgraph-checkpoint-postgres>=0.1.0

# LLM Providers
openai>=1.109.1

# Data processing
pandas>=2.0.0
numpy>=1.26.0
pydomo==0.3.0.13

# Azure
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.32.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
gunicorn>=21.2.0
